+++ core/src/main/java/org/ehcache/core/Ehcache.java
@@ -36,8 +36,8 @@ import org.ehcache.Status;
 import org.ehcache.config.CacheConfiguration;
 import org.ehcache.config.CacheRuntimeConfiguration;
 import org.ehcache.core.events.CacheEventDispatcher;
-import org.ehcache.core.resilience.LoggingRobustResilienceStrategy;
-import org.ehcache.core.resilience.RecoveryCache;
+import org.ehcache.core.internal.resilience.LoggingRobustResilienceStrategy;
+import org.ehcache.core.internal.resilience.RecoveryCache;
 import org.ehcache.core.spi.cache.Store;
 import org.ehcache.core.spi.cache.Store.ValueHolder;
 import org.ehcache.core.spi.cache.Store.PutStatus;
@@ -59,7 +59,7 @@ import org.ehcache.expiry.Duration;
 import org.ehcache.function.BiFunction;
 import org.ehcache.function.Function;
 import org.ehcache.function.NullaryFunction;
-import org.ehcache.resilience.ResilienceStrategy;
+import org.ehcache.core.internal.resilience.ResilienceStrategy;
 import org.ehcache.spi.LifeCycled;
 import org.ehcache.spi.loaderwriter.CacheLoaderWriter;
 import org.ehcache.statistics.BulkOps;
