+++ core/src/main/java/org/ehcache/core/EhcacheWithLoaderWriter.java
@@ -31,9 +31,9 @@ import org.ehcache.expiry.Duration;
 import org.ehcache.function.BiFunction;
 import org.ehcache.function.Function;
 import org.ehcache.function.NullaryFunction;
-import org.ehcache.core.resilience.LoggingRobustResilienceStrategy;
-import org.ehcache.core.resilience.RecoveryCache;
-import org.ehcache.resilience.ResilienceStrategy;
+import org.ehcache.core.internal.resilience.LoggingRobustResilienceStrategy;
+import org.ehcache.core.internal.resilience.RecoveryCache;
+import org.ehcache.core.internal.resilience.ResilienceStrategy;
 import org.ehcache.spi.LifeCycled;
 import org.ehcache.core.spi.cache.Store;
 import org.ehcache.core.spi.cache.Store.ValueHolder;
